Centralized Maven test utils methods
