+++ kie-wb-common-services/kie-wb-common-compiler/kie-wb-common-compiler-core/src/test/java/org/kie/workbench/common/services/backend/compiler/BaseCompilerTest.java
@@ -48,9 +48,9 @@ public class BaseCompilerTest {
 
     public BaseCompilerTest(String prjName) {
         try {
-            mavenRepo = TestUtil.getMavenRepo();
+            mavenRepo = TestUtilMaven.getMavenRepo();
             tmpRoot = Files.createTempDirectory("repo");
-            alternateSettingsAbsPath = TestUtil.getSettingsFile();
+            alternateSettingsAbsPath = TestUtilMaven.getSettingsFile();
             Path tmp = TestUtil.createAndCopyToDirectory(tmpRoot, "dummy", prjName);
             info = new WorkspaceCompilationInfo(Paths.get(tmp.toUri()));
         } catch (Exception e) {
