+++ kie-wb-common-services/kie-wb-common-compiler/kie-wb-common-compiler-core/src/test/java/org/kie/workbench/common/services/backend/utils/TestUtil.java
@@ -37,7 +37,6 @@ import org.uberfire.java.nio.file.Paths;
 public class TestUtil {
 
     private static Logger logger = LoggerFactory.getLogger(TestUtil.class);
-    private static final String JENKINS_SETTINGS_XML_FILE = "JENKINS_SETTINGS_XML_FILE";
 
     public static void copyTree(Path source,
                                 Path target) throws IOException {
@@ -89,48 +88,12 @@ public class TestUtil {
         }
     }
 
-    public static String getMavenRepo() throws Exception {
-        List<String> repos = Arrays.asList("M2_REPO", "MAVEN_REPO_LOCAL", "MAVEN_REPO", "M2_REPO_LOCAL");
-        String mavenRepo = "";
-        for (String repo : repos) {
-            if (System.getenv(repo) != null) {
-                mavenRepo = System.getenv(repo);
-                break;
-            }
-        }
-        return StringUtils.isEmpty(mavenRepo) ? createMavenRepo().toAbsolutePath().toString() : mavenRepo;
-    }
-
-    public static Path createMavenRepo() throws Exception {
-        Path mavenRepository = Paths.get(System.getProperty("user.home"),
-                                         "/.m2/repository");
-        if (!Files.exists(mavenRepository)) {
-            logger.info("Creating a m2_repo into " + mavenRepository);
-            if (!Files.exists(Files.createDirectories(mavenRepository))) {
-                logger.error("Folder not writable to create Maven repo{}", mavenRepository);
-                throw new Exception("Folder not writable to create Maven repo:"+mavenRepository);
-            }
-        }
-        return mavenRepository;
-    }
-
     public static Path createAndCopyToDirectory(Path root, String dirName, String copyTree) throws IOException {
         //NIO creation and copy content
         Path dir = Files.createDirectories(Paths.get(root.toString(), dirName));
         TestUtil.copyTree(Paths.get(copyTree), dir);
         return dir;
         //end NIO
-    }
-
-    public static String getSettingsFile(){
-        String jenkinsFile = System.getenv().get(JENKINS_SETTINGS_XML_FILE);
-        if(jenkinsFile != null){
-            logger.info("Using settings.xml file provided by JENKINS:{}", jenkinsFile);
-            return jenkinsFile;
-        }else {
-            logger.info("Using local settings.xml file.");
-            return new File("src/test/settings.xml").getAbsolutePath();
-        }
     }
 
     public static void saveMavenLogIfCompilationResponseNotSuccessfull(Path tmp, CompilationResponse response, Class<?> testClass, TestName testName) throws Exception{
