+++ kie-wb-common-services/kie-wb-common-compiler/kie-wb-common-compiler-testing/src/main/resources/logback.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration debug="true">
+
+  <appender name="consoleAppender" class="ch.qos.logback.core.ConsoleAppender">
+    <encoder>
+      <!-- %l lowers performance -->
+      <!--<pattern>%d [%t] %-5p %l%n  %m%n</pattern>-->
+      <!--<pattern>%d [%t] %-5p %m%n</pattern>-->
+      <pattern>%d [%thread] %level %logger{35} - %msg%n</pattern>
+    </encoder>
+  </appender>
+
+  <appender name="KieSift" class="org.kie.workbench.common.services.backend.logback.appender.KieSiftingAppender">
+    <discriminator class="org.kie.workbench.common.services.backend.logback.appender.UUIDThreadNameDiscriminator"/>
+    <sift>
+      <appender name="${compilation.ID}" class="org.kie.workbench.common.services.backend.logback.appender.KieSiftingAppender">
+        <encoder>
+          <pattern>%d [%thread] %level %logger{35} - %msg%n</pattern>
+        </encoder>
+      </appender>
+    </sift>
+  </appender>
+
+  <logger name="org.kie.workbench.common.services.backend.compiler" level="INFO" additivity="false">
+    <appender-ref ref="KieSift"/>
+    <appender-ref ref="consoleAppender"/>
+  </logger>
+
+  <!--<logger name="org.apache.maven" level="INFO" additivity="false">
+    <appender-ref ref="KieSift"/>
+  </logger>-->
+
+  <logger name="org.apache.maven" level="INFO" additivity="false">
+    <appender-ref ref="KieSift"/>
+    <appender-ref ref="consoleAppender"/>
+  </logger>
+
+  <logger name="org.drools" level="INFO"/>
+  <logger name="org.kie" level="INFO"/>
+  <logger name="org.guvnor" level="INFO"/>
+
+  <root level="INFO">
+    <appender-ref ref="KieSift"/>
+    <appender-ref ref="consoleAppender"/>
+  </root>
+
+</configuration>
